requirements.txt
roboticstoolbox-python>=0.11.0
swift-simulator
spatialmath-python>=0.9.0
spatialgeometry
numpy>=1.20.0
matplotlib>=3.3.0
qpsolvers>=1.8.0